requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.5
beautifulsoup4==4.12.3
APScheduler==3.10.4
Flask==3.0.3
